+++ x/staking/keeper/grpc_query_test.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/staking/teststaking"
+	"github.com/cosmos/cosmos-sdk/x/staking/testutil"
 	"github.com/cosmos/cosmos-sdk/x/staking/types"
 )
 
@@ -13,7 +13,7 @@ func (s *KeeperTestSuite) TestGRPCQueryValidator() {
 	ctx, keeper, queryClient := s.ctx, s.stakingKeeper, s.queryClient
 	require := s.Require()
 
-	validator := teststaking.NewValidator(s.T(), sdk.ValAddress(PKs[0].Address().Bytes()), PKs[0])
+	validator := testutil.NewValidator(s.T(), sdk.ValAddress(PKs[0].Address().Bytes()), PKs[0])
 	keeper.SetValidator(ctx, validator)
 	var req *types.QueryValidatorRequest
 	testCases := []struct {